src/get-data/requirements.txt
requests
alive_progress
pandas
psycopg2_binary
sqlalchemy
python-dotenv
